Flyttet og renamet filene for lokal kjøring
